MP: my first commit and its a big one!
I'm updating in one chunk as opposed to a series of functional updates in case there is a problem and we need to pull this version back out.

Here are the updates I've added:

DHCP Pool maint abilities.
DHCP Pool data is displayed on subnets and calculated in usages.
DHCP option maint abilities.
DHCP option data is displayed on subnets and partialy on hosts.
DHCP failover group maintinance (still needs work on "server" relations, this is mostly display purposes for now)
Other minor cleanup as I ran accross it.