MP: lots of fixes.  completed dhcp_server related code for display_subnets. changed the primary_dns record indicator from the interface list to the record list.  this is really where it should have been. html_desktop got a style sheet change so the screen looks a bit better (I hope)
I am starting to see some interesting things with the array_keys for the htmlentities lines.  Not sure why I'm seeing this now other than I just went through a php version change from 5 to 4.something.  I'm fixing it by casting the arrays: array_keys((array)$variable).  This should probably be done anyway.  It also seems to happen because in some cases the value of the array is empty.