MP: added permission checks to the main "tools" dropdown tooltip menu.
Also fixes for dns records.   now when you change an IP that an A record uses, the child records update to the new interface properly.  Also TXT records no longer require an interface since they are one of the few record types that does not actually need an IP associated.