Allow for MX,NS, and CNAME records to point to externals domains. Also allow for duplicate A/AAAA duplicate records.
Allowing MX, NS and CNAME records to point to external domains is needed
for example, for Gmail/GSuite MX Records. Also, if having other 3rd party
cloud services, this is sometimes needed.  Footer files are a solution,
however not for BIND-DLZ server, and for single source of truth.

This is an optional config, and when not enabled should have a NOOP effect:

    "allow_external_pointsto" => 1

HOWEVER - this does slightly break build-bind, and I have not yet provided
patches.

Allow Duplicate A,AAA records for the poor man's loadbalanceing with
Round-Robin DNS. This too could be done with footers, however, in the
spirit of a single source of truth - allowing this to be in the database
as well as allowing BIND-DLZ to server it.

    "allow_duplicate_arecords" => 1