fix custom db session handlers for changes in PHP7.1
As of PHP 7.1, $_SESSION is no longer intialized if the custom session
read handler returns false or null. With $_SEESION not initialized, the
the custom session write handler is not getting called (just close()).

The way the ONA customer session handlers are written, this new PHP behavior
results new sessions never being written to the database.